requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
requests==2.31.0
email-validator==2.1.0
python-dotenv==1.0.0
rich==13.7.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2
pytest==7.4.4
pytest-asyncio==0.21.1
streamlit==1.29.0
aiohttp==3.9.1
pandas==2.1.4